Refactored naming of auto-generated file names
