+++ frontend_scv/lib/screens/evidence_screen.dart
@@ -2,7 +2,7 @@ import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
 import 'package:unison/models/evidenceData.dart';
-import 'package:unison/widgets/agreement/evidenceListScreen.dart';
+import 'package:unison/widgets/agreement/evidence_list_panel.dart';
 import 'package:unison/widgets/agreement/evidence_upload_panel.dart';
 import 'package:unison/widgets/miscellaneous/funky_text_widget.dart';
 import 'package:mime/mime.dart';
@@ -24,12 +24,10 @@ class _MessagingScreenState extends State<EvidenceScreen> {
 
     List<EvidenceData> evidences = [];
 
-
     return Scaffold(
       appBar: AppBar(
         title: FunkyText('Evidence for Agreement ' + agreementId),
       ),
-
       body: EvidenceListPanel(agreementId),
     );
   }
