+++ frontend_scv/lib/widgets/agreement/evidence_display_item.dart
