+++ frontend_scv/lib/widgets/agreement/evidence_list_panel.dart
@@ -8,34 +8,32 @@ import 'package:unison/models/evidence.dart';
 import 'package:unison/models/evidenceData.dart';
 import 'package:unison/services/Server/evidenceService.dart';
 
-import 'EvidenceDisplayItem.dart';
+import 'evidence_display_item.dart';
 
 class EvidenceListPanel extends StatefulWidget {
-
   final String agreementId;
   EvidenceListPanel(this.agreementId);
 
   @override
   _EvidenceListPanelState createState() => _EvidenceListPanelState();
-
 }
 
 class _EvidenceListPanelState extends State<EvidenceListPanel> {
-
   TextEditingController _urlLinkController = new TextEditingController();
   EvidenceService evServe = EvidenceService();
 
   Future<void> uploadEvidence() async {
-      FilePickerResult picked = await FilePicker.platform.pickFiles();
+    FilePickerResult picked = await FilePicker.platform.pickFiles();
 
-      List<PlatformFile> fileRes = picked.files;
-      Evidence evid = Evidence(widget.agreementId);
-      await evid.setFile(fileRes[0]); //May be expanded in the future to allow multiple files
+    List<PlatformFile> fileRes = picked.files;
+    Evidence evid = Evidence(widget.agreementId);
+    await evid.setFile(
+        fileRes[0]); //May be expanded in the future to allow multiple files
 
-      await evServe.storeEvidence(evid);
-      setState(() {
-        build(context);
-      });
+    await evServe.storeEvidence(evid);
+    setState(() {
+      build(context);
+    });
   }
 
   Future<void> linkEvidence() async {
@@ -55,7 +53,8 @@ class _EvidenceListPanelState extends State<EvidenceListPanel> {
               TextButton(
                 child: Text('Save'),
                 onPressed: () async {
-                  await evServe.linkEvidence(_urlLinkController.text, widget.agreementId);
+                  await evServe.linkEvidence(
+                      _urlLinkController.text, widget.agreementId);
 
                   _urlLinkController.clear();
                   Navigator.of(context).pop();
@@ -84,31 +83,46 @@ class _EvidenceListPanelState extends State<EvidenceListPanel> {
   Widget build(BuildContext context) {
     return Scaffold(
       body: Container(
-
         child: FutureBuilder(
-            future: evServe.getEvidenceData(widget.agreementId), builder: (context, snapshot) {
-              List<Widget> cards = [Center(
-                child: Row(mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Center(child: TextButton(child: Text('Upload Evidence to UniServer(TM)'), onPressed: uploadEvidence,)),
-                    SizedBox(width: 100),
-                    Center(child: TextButton(child: Text('Provide a link to evidence (url)'), onPressed: linkEvidence,)),
-                  ],
-                ),
-              )];
+            future: evServe.getEvidenceData(widget.agreementId),
+            builder: (context, snapshot) {
+              List<Widget> cards = [
+                Center(
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Center(
+                          child: TextButton(
+                        child: Text('Upload Evidence to UniServer(TM)'),
+                        onPressed: uploadEvidence,
+                      )),
+                      SizedBox(width: 100),
+                      Center(
+                          child: TextButton(
+                        child: Text('Provide a link to evidence (url)'),
+                        onPressed: linkEvidence,
+                      )),
+                    ],
+                  ),
+                )
+              ];
               if (snapshot.connectionState != ConnectionState.done) {
-               return AwesomeLoader();
+                return AwesomeLoader();
               }
 
               List<EvidenceData> evs = snapshot.data;
-              for (int i =0;i<evs.length;i++) {
-                cards.add(SizedBox(height: 10,));
+              for (int i = 0; i < evs.length; i++) {
+                cards.add(SizedBox(
+                  height: 10,
+                ));
                 cards.add(EvidenceDisplayItem(evs[i], widget.agreementId));
               }
-              return SingleChildScrollView(child: Column(children: cards,));
-        }),
+              return SingleChildScrollView(
+                  child: Column(
+                children: cards,
+              ));
+            }),
       ),
     );
   }
-
 }
