+++ frontend_scv/lib/widgets/jury/judge_detail_info_panel.dart
